+++ app/controllers/data_sheets_controller.rb
@@ -2,12 +2,22 @@ class DataSheetsController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    @data_sheets = DataSheet.where(company_id: params[:company_id])
+    if CompanyUser.exists?(company_id: params[:company_id], user_id: current_user) || current_user.role == 'owner'
+      @data_sheets = DataSheet.where(company_id: params[:company_id])
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def new
-    @company = Company.find(params[:company_id])
-    @data_sheet = DataSheet.new
+    if CompanyUser.exists?(company_id: params[:company_id], user_id: current_user) && current_user.role == 'admin'
+      @company = Company.find(params[:company_id])
+      @data_sheet = DataSheet.new
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def create
@@ -23,8 +33,13 @@ class DataSheetsController < ApplicationController
   end
 
   def edit
-    @company = Company.find(params[:company_id])
-    @data_sheet = DataSheet.find(params[:id])
+    if (CompanyUser.exists?(company_id: params[:company_id], user_id: current_user) && current_user.role == 'admin') || current_user.role == 'owner'
+      @company = Company.find(params[:company_id])
+      @data_sheet = DataSheet.find(params[:id])
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   def update
@@ -40,14 +55,19 @@ class DataSheetsController < ApplicationController
   end
 
   def destroy
-    data_sheet = DataSheet.find(params[:id])
-    data_sheet.destroy
-    flash[:notice] = 'Data sheet deleted.'
-    redirect_to company_data_sheets_path
+    if (CompanyUser.exists?(company_id: params[:company_id], user_id: current_user) && current_user.role == 'admin') || current_user.role == 'owner'
+      data_sheet = DataSheet.find(params[:id])
+      data_sheet.destroy
+      flash[:notice] = 'Data sheet deleted.'
+      redirect_to company_data_sheets_path
+    else
+      flash[:alert] = "You don't have access to this page!"
+      redirect_to root_path
+    end
   end
 
   protected
   def data_sheet_params
-    params.require(:data_sheet).permit(:name, :description, :sds)
+    params.require(:data_sheet).permit(:number, :name, :sds)
   end
 end
